setup/cnec-discovery/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
apscheduler>=3.10.0
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.9.0
google-generativeai>=0.8.0
jinja2>=3.1.0
boto3>=1.34.0
zerobounce-sdk>=0.3.0
apify-client>=1.6.0
